+++ src/MapParallax.cpp
@@ -121,8 +121,8 @@ void MapParallax::render(const FPoint& cam, const std::string& map_layer) {
 		center_tile.y -= height/2;
 
 		Point draw_pos;
-		draw_pos.x = center_tile.x - static_cast<int>(ceil(static_cast<float>(VIEW_W_HALF + center_tile.x) / static_cast<float>(width))) * width;
-		draw_pos.y = center_tile.y - static_cast<int>(ceil(static_cast<float>(VIEW_H_HALF + center_tile.y) / static_cast<float>(height))) * height;
+		draw_pos.x = center_tile.x - static_cast<int>(ceilf(static_cast<float>(VIEW_W_HALF + center_tile.x) / static_cast<float>(width))) * width;
+		draw_pos.y = center_tile.y - static_cast<int>(ceilf(static_cast<float>(VIEW_H_HALF + center_tile.y) / static_cast<float>(height))) * height;
 		Point start_pos = draw_pos;
 
 		while (draw_pos.x < VIEW_W) {
